Reuse Favicon component and move inside PageHeader
