+++ common/views/components/PageLayout/index.tsx
@@ -39,6 +39,8 @@ import { JsonLdObj } from '@weco/common/views/components/JsonLd';
 import NewsletterPromo from '@weco/common/views/components/NewsletterPromo';
 import PopupDialog from '@weco/common/views/components/PopupDialog';
 
+import Favicons from './PageLayout.Favicons';
+
 type HeaderProps = {
   customNavLinks: NavLink[];
   isMinimalHeader?: boolean;
@@ -213,37 +215,8 @@ const PageLayoutComponent: FunctionComponent<Props> = ({
           name="viewport"
           content="width=device-width, initial-scale=1"
         />
-        <link
-          rel="apple-touch-icon"
-          sizes="180x180"
-          href="https://i.wellcomecollection.org/assets/icons/apple-touch-icon.png"
-        />
-        <link
-          rel="shortcut icon"
-          href="https://i.wellcomecollection.org/assets/icons/favicon.ico"
-          type="image/ico"
-        />
-        <link
-          rel="icon"
-          type="image/png"
-          href="https://i.wellcomecollection.org/assets/icons/favicon-32x32.png"
-          sizes="32x32"
-        />
-        <link
-          rel="icon"
-          type="image/png"
-          href="https://i.wellcomecollection.org/assets/icons/favicon-16x16.png"
-          sizes="16x16"
-        />
-        <link
-          rel="manifest"
-          href="https://i.wellcomecollection.org/assets/icons/manifest.json"
-        />
-        <link
-          rel="mask-icon"
-          href="https://i.wellcomecollection.org/assets/icons/safari-pinned-tab.svg"
-          color="#000000"
-        />
+
+        <Favicons />
 
         {rssUrl && (
           <link
@@ -364,3 +337,4 @@ const PageLayout: FunctionComponent<Props> = (props: Props) => {
 };
 
 export default PageLayout;
+export { Favicons };
