+++ identity/webapp/components/PageWrapper/index.tsx
@@ -6,9 +6,9 @@ import { useAppContext } from '@weco/common/contexts/AppContext';
 import useIsFontsLoaded from '@weco/common/hooks/useIsFontsLoaded';
 import { usePrismicData } from '@weco/common/server-data/Context';
 import { transformCollectionVenues } from '@weco/common/services/prismic/transformers/collection-venues';
-import Favicons from '@weco/common/views/components/Favicons';
 import Footer from '@weco/common/views/components/Footer';
 import Header from '@weco/common/views/components/Header';
+import { Favicons } from '@weco/common/views/components/PageLayout';
 import { GlobalStyle } from '@weco/common/views/themes/default';
 
 const Main = styled.div`
